Added socket timeout to arguments
Break out the timeout (a timeout results in OSError: [Errno 116] ETIMEDOUT), giving the user the option of setting the value of seconds, and using try/except to catch this. The alternative is to leave it as 'None', which would block forever (probably not the desired behaviour for embedded projects!) Note: Using a value of '0' (non-blocking) appears to trigger the EAGAIN error.